[REFACTOR] 유튜브 추천 작업 producer 패키지 변경
